refactor(portfolio): Created a Home.tsx file and transferred the landing page code to there
