+++ src/App.tsx
@@ -1,23 +1,8 @@
 import React from "react";
-import ThemeSwitcher from "./ThemeSwitcher.js";
+import Home from "./pages/Home.js";
 
 const App = () => {
-	return (
-		<>
-			<div className="absolute top-4 right-4">
-				<ThemeSwitcher />
-			</div>
-
-			<div className="flex flex-col items-center justify-center min-h-screen gap-1">
-				<h1 className="text-2xl font-family-heading text-primary-accent">
-					Francisca Portugal
-				</h1>
-				<p className="text-base font-family-body text-primary">
-					Web developer!
-				</p>
-			</div>
-		</>
-	);
+	return <Home />;
 };
 
 export default App;
