+++ src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "./style/index.css";
 import App from "./App.js";
-import ThemeProvider from "./ThemeProvider.js";
+import ThemeProvider from "./components/ThemeProvider.js";
 
 const rootElement = document.getElementById("root");
 if (!rootElement) throw new Error("Root element not found");
